Moved projectDTO to response folder
