+++ src/main/java/com/vena_project/crowd_funding/service/AdminService.java
@@ -1,6 +1,5 @@
 package com.vena_project.crowd_funding.service;
 
-import com.vena_project.crowd_funding.dto.ProjectDTO;
 import com.vena_project.crowd_funding.dto.ResponseDTO.UserResponseDTO;
 import com.vena_project.crowd_funding.model.Project;
 import com.vena_project.crowd_funding.model.enums.ProjectStatus;
