+++ src/main/java/com/vena_project/crowd_funding/service/ProjectService.java
@@ -1,7 +1,7 @@
 package com.vena_project.crowd_funding.service;
 
 import com.vena_project.crowd_funding.dto.RequestDTO.ProjectRequestDTO;
-import com.vena_project.crowd_funding.dto.ProjectDTO;
+import com.vena_project.crowd_funding.dto.ResponseDTO.ProjectDTO;
 import com.vena_project.crowd_funding.dto.ResponseDTO.ProjectResponseDTO;
 import com.vena_project.crowd_funding.model.Project;
 import com.vena_project.crowd_funding.model.enums.ProjectStatus;
