+++ src/main/java/com/vena_project/crowd_funding/service/impl/ProjectServiceImpl.java
@@ -2,7 +2,7 @@ package com.vena_project.crowd_funding.service.impl;
 
 import com.vena_project.crowd_funding.dto.RequestDTO.ProjectRequestDTO;
 import com.vena_project.crowd_funding.dto.ResponseDTO.ProjectResponseDTO;
-import com.vena_project.crowd_funding.dto.ProjectDTO;
+import com.vena_project.crowd_funding.dto.ResponseDTO.ProjectDTO;
 import com.vena_project.crowd_funding.exception.CreateAccessException;
 import com.vena_project.crowd_funding.exception.ResourceNotFoundException;
 import com.vena_project.crowd_funding.model.Project;
@@ -12,7 +12,6 @@ import com.vena_project.crowd_funding.model.enums.UserRole;
 import com.vena_project.crowd_funding.repository.ProjectRepository;
 import com.vena_project.crowd_funding.service.ProjectService;
 import com.vena_project.crowd_funding.service.UserService;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import java.time.LocalDate;
 import java.util.List;
